mdq/pyproject.toml
[project]
name = "mdq"
version = "0.1.0"
description = "A Markdown/JSON Question format for the Codehood project."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "linkify-it-py>=2.0.3",
    "markdown-it-py[plugins]>=4.0.0",
    "pydantic>=2.11.10",
    "rich>=14.1.0",
]
[project.optional-dependencies]
dev = []

[project.scripts]
mdq = "mdq.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.4.2"]
